add basic substitution optimizations
in these optimizations, some common mcl patterns observed in my own
programs are replaced by other, faster instructions.
for example, the pattern

    push 1
    add

is practically an increment, so a new instruction `inc` was added. you
can use either `inc` or the above, as both will now result in the same
code.

    push 1
    sub

is a similar pattern, replaced by the new instruction `dec`.

    over
    over

serves as a pattern for duplicating the current two top stack elements,
so effectively a dup, but for two at a time. accordingly, this is now
replaced by a `dup2` instruction by the optimization step of the
compiler.